Create 227. Basic Calculator II.py
This is a classic stack problem. We need to build a calculator, which means that timing and dividing has higher priority that adding and minusing. To achieve this, we have to go further to check the operation and then calcutate. Once it is done, we need to add the elements together because there may be some elements with lower pirority left.