Create 224. Basic Calculator.py
This is another problem using stack and in this problem we do not have to think about *or/ but we have to calculate the part inside the () first. 
We traverse though the string and calculate the digit directly. When we find +- we just change the signal and multiple with the res. 
Once we meet (, we have to append the res and signal into the stack and release when we find )