Create 456. 132 Pattern.py
This is a very tricky question for stack. Because when we are trying to traverse the list from beginning, we have to find a value between the max and min, and at the same time the index of mid should be bigger than the index of max and min. So we can go backwards, to traverse from the end. Because we can first get the third number and find the bigger and smaller one, which would be much easier to solve.